Create new-ann-bar.liquid
Go to your theme, edit the code, add this new file in Sections, name it whatever you want I named it new ann bar, add the code and save it. 
Go to customize, go to announcement, hide it, and you will also see a new type of announcement, simply click on it and start playing around with it.
This will make your bar slide, and have more freedom to adjust the font and height etc.